requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
cssselect2==0.7.0
Cython==3.0.8
defusedxml==0.7.1
diff-match-patch==20230430
dj-user-login-history==1.0.5
Django==5.0.1
django-admin-rangefilter==0.12.0
django-humanize==0.1.2
django-import-export==3.3.3
django-leaflet==0.28.3
django-semantic-admin==0.4.0
et-xmlfile==1.1.0
GDAL==3.4.3
geojson==3.0.1
gunicorn==21.2.0
html5lib==1.1
humanize==4.9.0
idna==3.6
joblib==1.3.2
lxml==5.1.0
MarkupPy==1.14
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==2.0.1
patsy==0.5.6
pillow==10.2.0
pmdarima==2.0.3
psycopg2-binary==2.9.6
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.4
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.1
svglib==1.5.1
tablib==3.5.0
threadpoolctl==3.2.0
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
